Refactor UI components: migrate Input, Badge, and Card components to Student directory
